- Rename `godot-core::export` to `godot-core::property` - move `godot-macros/src/derive_godot_class.rs` and `godot-macros/src/util.rs` into their own folders
